[REF] Port forwarding of the profiler buttons
