+++ profiler/static/src/less/player.less
@@ -1,0 +1,24 @@
+#oe_main_menu_navbar .o_menu_systray {
+
+    .oe_topbar_item.profiler_player i {
+        cursor: pointer;
+    }
+    .oe_topbar_item.profiler_player.profiler_player_enabled a.profiler_enable {
+        display: None;
+    }
+    .oe_topbar_item.profiler_player.profiler_player_enabled a.profiler_dump {
+        display: None;
+    }
+    .oe_topbar_item.profiler_player.profiler_player_disabled a.profiler_disable {
+        display: None;
+    }
+    .oe_topbar_item.profiler_player.profiler_player_clear a.profiler_disable {
+        display: None;
+    }
+    .oe_topbar_item.profiler_player.profiler_player_clear a.profiler_clear {
+        display: None;
+    }
+    .oe_topbar_item.profiler_player.profiler_player_clear a.profiler_dump {
+        display: None;
+    }
+}
